[test] Make calculateTrip more atomic
`calculateTrip()` was not doing *one thing*. This method was actually
creating vehicle and planetary body objects and *then* passing these
objects to another method `doActualCalculations()`. The fact that we
have nested a method within `calculateTrip` to *actually* do a
calculation means we are technically not doing one thing.
`calculateTrip()` has been decoupled to the following two methods:

- `calculateTrip` and `calculateNextTrip()` have been named to
`getInitialDialogOptions()` and `getNextDialogOptions()` and now return
the dialog options in an ArrayList<Object>
- `doActualCalculations` has been renamed to `calculateTripData()` and
  accepts the ArrayList<Object> of dialog options form the previous
  methods.